salman/pyproject.toml
[tool.poetry]
name = "salman"
version = "0.1.0"
description = ""
authors = ["Yiorgis Gozadinos <ggozadinos@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
langchain = "^0.0.135"
python-dotenv = "^1.0.0"
typer = {version = "^0.7.0", extras = ["all"]}
anthropic = "^0.2.6"
google-search-results = "^2.4.2"
llvmlite = "^0.39.1"
torch = "1.13.1"
numba = "^0.56.4"
torchaudio = "0.13.1"
openai-whisper = {git = "https://github.com/openai/whisper.git"}
pydub = "^0.25.1"
hmmlearn = {git = "https://github.com/hmmlearn/hmmlearn.git"}
pyannote-audio = "^2.1.1"


[tool.poetry.group.dev.dependencies]
pdbpp = "^0.10.3"
ruff = "^0.0.261"
black = "^23.3.0"
pytest = "^7.3.0"

[tool.poetry.scripts]
prompt = "salman.cli.prompt:app"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
addopts = "-rP --disable-warnings"
testpaths = [
    "tests",
]

[tool.black]
line-length = 88

[tool.ruff]

line-length = 88

# Enable Flake's "E" and "F" codes by default and "I" for sorting imports.
select = ["E", "F", "I"]
ignore = ["E501", "E741"] # E741 should not be ignored

# Exclude a variety of commonly ignored directories.
exclude = [
    ".direnv",
    ".eggs",
    ".git",
    ".mypy_cache",
    ".nox",
    ".ruff_cache",
    ".tox",
    ".venv",
    "__pypackages__",
    ".cache",
    "_build",
    "buck-out",
    "build",
    "dist",
    "node_modules",
    "venv",
]
per-file-ignores = {"__init__.py"=["F401","F403"]}

# Allow unused variables when underscore-prefixed.
dummy-variable-rgx = "^(_+|(_+[a-zA-Z0-9_]*[a-zA-Z0-9]+?))$"

# Assume Python 3.10.
target-version = "py310"
